Install Miniconda into a no-capitalization directory name
It turns out that conda-build's detection of the build prefix is
case-sensitive on Windows, whereas CMake likes to de-capitalize any paths that
it handles. So when building packages in a directory with capital letters like
C:\Tools\Miniconda3, you would end up with packages where the detection of the
build prefix failed. Here we change things to install into C:\mc3, which
should avoid the problem. A more thoroughgoing solution would be nice but this
seems to be what has worked well for Anaconda, Inc.